Abstract and clean up
